drivers/gpu/drm/amd/pm: Fix possible null pointer dereference in `smu_v11_0_get_fan_speed_pwm()`
The `speed` is assumed to be not null in
`smu_v11_0_get_fan_speed_pwm()` but in
`drivers/gpu/drm/amd/pm/amdgpu_pm.c:3427`, it is
called with `NULL`.  The implementation is used by
Navi 10 and Sienna GPUs. For some attributes (like
`sensor_dev_attr_temp1_input.dev_attr.attr`), the
function `amdgpu_dpm_get_fan_speed_pwm` with a
`NULL` speed.  The function is stored in an
`attribute_group` and can be executed.

Fixes: fb1f667e71c0 ("drm/amd/pm: correct the fan speed PWM retrieving")
Signed-off-by: Thomas Fourier <fourier.thomas@gmail.com>